+++ Sources/Mixpanel.swift
@@ -25,7 +25,8 @@ open class Mixpanel {
      - parameter token:                     your project token
      - parameter trackAutomaticEvents:      Whether or not to collect common mobile events.
      - parameter flushInterval:             Optional. Interval to run background flushing
-     - parameter instanceName:              Optional. The name you want to call this instance, must be unique 1:1 for each instance's project token. Defaults to project token.
+     - parameter instanceName:              Optional. The name you want to uniquely identify the Mixpanel Instance.
+                                            It is useful when you want more than one Mixpanel instance under the same project token.
      - parameter optOutTrackingByDefault:   Optional. Whether or not to be opted out from tracking by default
      - parameter useUniqueDistinctId:       Optional. whether or not to use the unique device identifier as the distinct_id
      - parameter superProperties:           Optional. Super properties dictionary to register during initialization