+++ Sources/MixpanelInstance.swift
@@ -198,7 +198,6 @@ open class MixpanelInstance: CustomDebugStringConvertible, FlushDelegate, AEDele
         }
     }
 #endif // DECIDE
-    
     var superProperties = InternalProperties()
     var trackingQueue: DispatchQueue
     var networkQueue: DispatchQueue
@@ -235,11 +234,12 @@ open class MixpanelInstance: CustomDebugStringConvertible, FlushDelegate, AEDele
         let label = "com.mixpanel.\(self.apiToken)"
         trackingQueue = DispatchQueue(label: "\(label).tracking)", qos: .utility)
         networkQueue = DispatchQueue(label: "\(label).network)", qos: .utility)
-        mixpanelPersistence = MixpanelPersistence.init(token: self.apiToken)
+        self.name = name
+        
+        mixpanelPersistence = MixpanelPersistence.init(instanceName: name)
         mixpanelPersistence.migrate()
         self.useUniqueDistinctId = useUniqueDistinctId
         
-        self.name = name
         readWriteLock = ReadWriteLock(label: "com.mixpanel.globallock")
         flushInstance = Flush(basePathIdentifier: name)
 #if DECIDE
@@ -247,6 +247,7 @@ open class MixpanelInstance: CustomDebugStringConvertible, FlushDelegate, AEDele
 #endif // DECIDE
         sessionMetadata = SessionMetadata(trackingQueue: trackingQueue)
         trackInstance = Track(apiToken: self.apiToken,
+                              instanceName: self.name,
                               lock: self.readWriteLock,
                               metadata: sessionMetadata, mixpanelPersistence: mixpanelPersistence)
         trackInstance.mixpanelInstance = self
@@ -298,13 +299,13 @@ open class MixpanelInstance: CustomDebugStringConvertible, FlushDelegate, AEDele
         if trackAutomaticEvents {
             MixpanelPersistence.saveAutomaticEventsEnabledFlag(value: trackAutomaticEvents,
                                                                 fromDecide: false,
-                                                                apiToken: self.apiToken)
+                                                               apiToken: self.apiToken)
         }
         
 #if DECIDE || TV_AUTO_EVENTS
         if !MixpanelInstance.isiOSAppExtension() {
             automaticEvents.delegate = self
-            automaticEvents.initializeEvents(apiToken: self.apiToken)
+            automaticEvents.initializeEvents(instanceName: self.name)
         }
 #endif // DECIDE
     }
@@ -611,6 +612,12 @@ extension MixpanelInstance {
                 self.hadPersistedDistinctId = true
             }
             
+            if self.userId == nil {
+                self.readWriteLock.write {
+                    self.userId = distinctId
+                }
+            }
+            
             if distinctId != self.distinctId {
                 let oldDistinctId = self.distinctId
                 self.readWriteLock.write {
@@ -636,7 +643,7 @@ extension MixpanelInstance {
                 anonymousId: self.anonymousId,
                 userId: self.userId,
                 alias: self.alias,
-                hadPersistedDistinctId: self.hadPersistedDistinctId), apiToken: self.apiToken)
+                hadPersistedDistinctId: self.hadPersistedDistinctId), instanceName: self.name)
             if let completion = completion {
                 DispatchQueue.main.async(execute: completion)
             }
@@ -651,6 +658,9 @@ extension MixpanelInstance {
      The alias method creates an alias which Mixpanel will use to remap one id to another.
      Multiple aliases can point to the same identifier.
      
+     Please note: With Mixpanel Identity Merge enabled, calling alias is no longer required
+     but can be used to merge two IDs in scenarios where identify() would fail
+     
      
      `mixpanelInstance.createAlias("New ID", distinctId: mixpanelInstance.distinctId)`
      
@@ -662,10 +672,12 @@ extension MixpanelInstance {
      - parameter alias:      A unique identifier that you want to use as an identifier for this user.
      - parameter distinctId: The current user identifier.
      - parameter usePeople: boolean that controls whether or not to set the people distinctId to the event distinctId.
+     - parameter andIdentify: an optional boolean that controls whether or not to call 'identify' with your current
+     user identifier(not alias). Default to true for keeping your signup funnels working correctly in most cases.
      - parameter completion: an optional completion handler for when the createAlias has completed.
      This should only be set to false if you wish to prevent people profile updates for that user.
      */
-    open func createAlias(_ alias: String, distinctId: String, usePeople: Bool = true, completion: (() -> Void)? = nil) {
+    open func createAlias(_ alias: String, distinctId: String, usePeople: Bool = true, andIdentify: Bool = true, completion: (() -> Void)? = nil) {
         if hasOptedOutTracking() {
             if let completion = completion {
                 DispatchQueue.main.async(execute: completion)
@@ -725,12 +737,14 @@ extension MixpanelInstance {
                     anonymousId: anonymousIdSnapshot,
                     userId: userIdSnapshot,
                     alias: aliasSnapshot,
-                    hadPersistedDistinctId: hadPersistedDistinctIdSnapshot), apiToken: self.apiToken)
+                    hadPersistedDistinctId: hadPersistedDistinctIdSnapshot), instanceName: self.name)
             }
             
             let properties = ["distinct_id": distinctId, "alias": alias]
             track(event: "$create_alias", properties: properties)
-            identify(distinctId: distinctId, usePeople: usePeople)
+            if andIdentify {
+                identify(distinctId: distinctId, usePeople: usePeople)
+            }
             flush(completion: completion)
         } else {
             Logger.error(message: "alias: \(alias) matches distinctId: \(distinctId) - skipping api call.")
@@ -753,7 +767,7 @@ extension MixpanelInstance {
                 return
             }
             
-            MixpanelPersistence.deleteMPUserDefaultsData(apiToken: self.apiToken)
+            MixpanelPersistence.deleteMPUserDefaultsData(instanceName: self.name)
             self.readWriteLock.write {
                 self.timedEvents = InternalProperties()
                 self.distinctId = self.defaultDistinctId()
@@ -781,24 +795,24 @@ extension MixpanelInstance {
     
     open func archive() {
         self.readWriteLock.read {
-            MixpanelPersistence.saveTimedEvents(timedEvents: timedEvents, apiToken: apiToken)
-            MixpanelPersistence.saveSuperProperties(superProperties: superProperties, apiToken: apiToken)
+            MixpanelPersistence.saveTimedEvents(timedEvents: timedEvents, instanceName: self.name)
+            MixpanelPersistence.saveSuperProperties(superProperties: superProperties, instanceName: self.name)
             MixpanelPersistence.saveIdentity(MixpanelIdentity.init(
                 distinctID: distinctId,
                 peopleDistinctID: people.distinctId,
                 anonymousId: anonymousId,
                 userId: userId,
                 alias: alias,
-                hadPersistedDistinctId: hadPersistedDistinctId), apiToken: apiToken)
+                hadPersistedDistinctId: hadPersistedDistinctId), instanceName: self.name)
         }
     }
     
     func unarchive() {
         self.readWriteLock.write {
-            optOutStatus = MixpanelPersistence.loadOptOutStatusFlag(apiToken: apiToken)
-            superProperties = MixpanelPersistence.loadSuperProperties(apiToken: apiToken)
-            timedEvents = MixpanelPersistence.loadTimedEvents(apiToken: apiToken)
-            let mixpanelIdentity = MixpanelPersistence.loadIdentity(apiToken: apiToken)
+            optOutStatus = MixpanelPersistence.loadOptOutStatusFlag(instanceName: self.name)
+            superProperties = MixpanelPersistence.loadSuperProperties(instanceName: self.name)
+            timedEvents = MixpanelPersistence.loadTimedEvents(instanceName: self.name)
+            let mixpanelIdentity = MixpanelPersistence.loadIdentity(instanceName: self.name)
             (distinctId, people.distinctId, anonymousId, userId, alias, hadPersistedDistinctId) = (
                 mixpanelIdentity.distinctID,
                 mixpanelIdentity.peopleDistinctID,
@@ -818,7 +832,7 @@ extension MixpanelInstance {
                     anonymousId: anonymousId,
                     userId: userId,
                     alias: alias,
-                    hadPersistedDistinctId: hadPersistedDistinctId), apiToken: apiToken)
+                    hadPersistedDistinctId: hadPersistedDistinctId), instanceName: self.name)
             }
         }
     }
@@ -879,7 +893,7 @@ extension MixpanelInstance {
             // automatic events will NOT be flushed until one of the flags is non-nil
             let eventQueue = self.mixpanelPersistence.loadEntitiesInBatch(
                 type: self.persistenceTypeFromFlushType(.events),
-                excludeAutomaticEvents: !MixpanelPersistence.automaticEventsFlagIsSet(apiToken: self.apiToken)
+                excludeAutomaticEvents: !MixpanelPersistence.automaticEventsFlagIsSet(instanceName: self.name)
             )
             let peopleQueue = self.mixpanelPersistence.loadEntitiesInBatch(type: self.persistenceTypeFromFlushType(.people))
             let groupsQueue = self.mixpanelPersistence.loadEntitiesInBatch(type: self.persistenceTypeFromFlushType(.groups))
@@ -1107,7 +1121,7 @@ extension MixpanelInstance {
             self.readWriteLock.write {
                 self.timedEvents = timedEvents
             }
-            MixpanelPersistence.saveTimedEvents(timedEvents: timedEvents, apiToken: self.apiToken)
+            MixpanelPersistence.saveTimedEvents(timedEvents: timedEvents, instanceName: self.name)
         }
     }
     
@@ -1132,7 +1146,7 @@ extension MixpanelInstance {
             self.readWriteLock.write {
                 self.timedEvents = InternalProperties()
             }
-            MixpanelPersistence.saveTimedEvents(timedEvents: InternalProperties(), apiToken: self.apiToken)
+            MixpanelPersistence.saveTimedEvents(timedEvents: InternalProperties(), instanceName: self.name)
         }
     }
     
@@ -1146,7 +1160,7 @@ extension MixpanelInstance {
             guard let self = self else { return }
             
             let updatedTimedEvents = self.trackInstance.clearTimedEvent(event: event, timedEvents: self.timedEvents)
-            MixpanelPersistence.saveTimedEvents(timedEvents: updatedTimedEvents, apiToken: self.apiToken)
+            MixpanelPersistence.saveTimedEvents(timedEvents: updatedTimedEvents, instanceName: self.name)
         }
     }
     
@@ -1170,7 +1184,7 @@ extension MixpanelInstance {
         trackingQueue.async { [weak self] in
             guard let self = self else { return }
             self.superProperties = self.trackInstance.clearSuperProperties(self.superProperties)
-            MixpanelPersistence.saveSuperProperties(superProperties: self.superProperties, apiToken: self.apiToken)
+            MixpanelPersistence.saveSuperProperties(superProperties: self.superProperties, instanceName: self.name)
         }
     }
     
@@ -1194,7 +1208,7 @@ extension MixpanelInstance {
                 self.superProperties = updatedSuperProperties
             }
             self.readWriteLock.read {
-                MixpanelPersistence.saveSuperProperties(superProperties: self.superProperties, apiToken: self.apiToken)
+                MixpanelPersistence.saveSuperProperties(superProperties: self.superProperties, instanceName: self.name)
             }
         }
     }
@@ -1224,7 +1238,7 @@ extension MixpanelInstance {
                 self.superProperties = updatedSuperProperties
             }
             self.readWriteLock.read {
-                MixpanelPersistence.saveSuperProperties(superProperties: self.superProperties, apiToken: self.apiToken)
+                MixpanelPersistence.saveSuperProperties(superProperties: self.superProperties, instanceName: self.name)
             }
         }
     }
@@ -1247,7 +1261,7 @@ extension MixpanelInstance {
             guard let self = self else { return }
             self.superProperties = self.trackInstance.unregisterSuperProperty(propertyName,
                                                                               superProperties: self.superProperties)
-            MixpanelPersistence.saveSuperProperties(superProperties: self.superProperties, apiToken: self.apiToken)
+            MixpanelPersistence.saveSuperProperties(superProperties: self.superProperties, instanceName: self.name)
         }
     }
     
@@ -1263,7 +1277,7 @@ extension MixpanelInstance {
             self.trackInstance.updateSuperProperty(update,
                                                    superProperties: &superPropertiesShadow)
             self.superProperties = superPropertiesShadow
-            MixpanelPersistence.saveSuperProperties(superProperties: self.superProperties, apiToken: self.apiToken)
+            MixpanelPersistence.saveSuperProperties(superProperties: self.superProperties, instanceName: self.name)
         }
     }
     
@@ -1389,14 +1403,14 @@ extension MixpanelInstance {
                 self.anonymousId = self.distinctId
                 self.hadPersistedDistinctId = true
                 self.superProperties = InternalProperties()
-                MixpanelPersistence.saveTimedEvents(timedEvents: InternalProperties(), apiToken: self.apiToken)
+                MixpanelPersistence.saveTimedEvents(timedEvents: InternalProperties(), instanceName: self.name)
             }
             self.archive()
             self.readWriteLock.write {
                 self.optOutStatus = true
             }
             self.readWriteLock.read {
-                MixpanelPersistence.saveOptOutStatusFlag(value: self.optOutStatus!, apiToken: self.apiToken)
+                MixpanelPersistence.saveOptOutStatusFlag(value: self.optOutStatus!, instanceName: self.name)
             }
             
         }
@@ -1421,7 +1435,7 @@ extension MixpanelInstance {
                 self.optOutStatus = false
             }
             self.readWriteLock.read {
-                MixpanelPersistence.saveOptOutStatusFlag(value: self.optOutStatus!, apiToken: self.apiToken)
+                MixpanelPersistence.saveOptOutStatusFlag(value: self.optOutStatus!, instanceName: self.name)
             }
             if let distinctId = distinctId {
                 self.identify(distinctId: distinctId)
@@ -1475,7 +1489,7 @@ extension MixpanelInstance {
                                             token: self.apiToken)
             self.trackingQueue.async { [weak self] in
                 guard let self = self else { return }
-                if !MixpanelPersistence.loadAutomaticEventsEnabledFlag(apiToken: self.apiToken) {
+                if !MixpanelPersistence.loadAutomaticEventsEnabledFlag(instanceName: self.name) {
                     self.mixpanelPersistence.removeAutomaticEvents()
                 }
             }